requirements.txt
# Hub Dependencies

# ===== Web 框架 =====
fastapi>=0.100.0
uvicorn>=0.23.0

# ===== 数据验证 =====
pydantic[email]>=2.0.0

# ===== 认证 =====
PyJWT>=2.8.0

# ===== HTTP =====
httpx>=0.24.0
python-multipart>=0.0.6

# ===== 环境变量 =====
python-dotenv>=1.0.0
